lesson10/liyongli/devops/requirements.txt
Django
django-pure-pagination
mysqlclient
Pillow